COMP: incorporate the latest patches from VTK

+++ Wrapping/vtkParse.tab.c
@@ -182,8 +182,8 @@ int  HaveComment;
 char CommentText[50000];
 int CommentState;
 int sigClosed;
-int sigMark[10];
-int sigMarkDepth = 0;
+size_t sigMark[10];
+size_t sigMarkDepth = 0;
 unsigned int sigAllocatedLength;
 char *currentId = 0;
 
@@ -243,7 +243,7 @@ void preSig(const char *arg)
     }
   else if (!sigClosed)
     {
-    int m, n;
+    size_t m, n;
     char *cp;
     checkSigSize(arg);
     cp = currentFunction->Signature;
@@ -265,7 +265,7 @@ void postSig(const char *arg)
     }
   else if (!sigClosed)
     {
-    int m, n;
+    size_t m, n;
     char *cp;
     checkSigSize(arg);
     cp = currentFunction->Signature;
@@ -286,7 +286,7 @@ void preScopeSig(const char *arg)
     }
   else if (!sigClosed)
     {
-    int i, m, n, depth;
+    size_t i, m, n, depth;
     char *cp;
     checkSigSize(arg);
     cp = currentFunction->Signature;
@@ -363,7 +363,7 @@ void swapSig()
     }
   if (currentFunction->Signature && sigMark[sigMarkDepth] > 0)
     {
-    int i, m, n, nn;
+    size_t i, m, n, nn;
     char c;
     char *cp;
     cp = currentFunction->Signature;
@@ -392,7 +392,7 @@ void chopSig(void)
 {
   if (currentFunction->Signature)
     {
-    int n = strlen(currentFunction->Signature);
+    size_t n = strlen(currentFunction->Signature);
     if (n > 0)
       {
       currentFunction->Signature[n-1] = '\0';
